backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.3

# Phase 3: Email & Notifications
fastapi-mail==1.4.1
twilio==8.11.0
jinja2==3.1.3
